pyproject.toml
[project]
name = "mcp-github-issue-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.14.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
]
